[REFACTOR] 컨트롤러 모듈화 + 서비스계층 뼈대 작성
