+++ controllers/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  mainController: require('./main'),
+  testController: require('./test'),
+  userController: require('./user')
+};
