+++ routes/main.js
@@ -1,10 +1,9 @@
 const express = require('express');
 const router = express.Router();
 
-const mainController = require('../controllers/main');
+const {mainController} = require('../controllers');
 
 // 라우터 작성
 router.get('/', mainController.getTestsAndCategories);
 
 module.exports = router;
-
