+++ service/categoryService.js
@@ -1,0 +1,5 @@
+const {Category} = require('../models');
+
+module.exports = {
+  // Category 테이블에 접근하는 함수들 작성
+}
