+++ service/questionService.js
@@ -1,0 +1,5 @@
+const {Question} = require('../models');
+
+module.exports = {
+  // Question 테이블에 접근하는 함수들 작성
+}
